chore: get rid of critical security alert
the tooling and github are complaining about us using a version of lodash with
the worst possible security rating - for no good reason!

that version just happens to be in the package-lock.json of the packages we're
removing here, so there actually has not been a security issue as we use other
dependencies in the main bundle.

for peace of mind when looking at github and also for reasons of maintainability
i propose to get rid of the "packages", which are rather one-off factorings of
stuff that create a whole lot of noise, mental overhead and complex tooling.

goodbye little package!
