--- packages/cosmos-client/README.md
@@ -1,3 +1,0 @@
-# Cosmos Client
-
-This package provides a client for making requests to the Cosmos service of a DC/OS cluster.
