--- packages/cosmos-client/index.ts
@@ -1,7 +1,0 @@
-import { CosmosClient } from "./src/CosmosClient";
-
-export { CosmosClient, PackageVersionsResponse };
-
-interface PackageVersionsResponse {
-  results: Record<string, string>;
-}
